Make MinimalVOlume work correctly
Also change all floats to doubles.

+++ include/beam_graph.hpp
@@ -44,7 +44,7 @@ class BeamGraph{
     ProjectData * const data;
     std::vector<BeamNode*> nodes;
     BeamElementFactory::BeamElementType type;
-    void insert_element_matrix(std::vector<float>& K, const std::vector<float>& k, const std::vector<long>& pos, int w, int& n) const;
+    void insert_element_matrix(std::vector<double>& K, const std::vector<double>& k, const std::vector<long>& pos, int w, int& n) const;
     void clear_nodes();
 };
