+++ include/finite_element/direct_solver.hpp
@@ -27,10 +27,10 @@ namespace finite_element{
 
 class DirectSolver : public FiniteElement{
     public:
-    virtual std::vector<float> calculate_displacements(ProjectData* data, Meshing* mesh, const std::vector<double>& density = std::vector<double>(), const std::vector<float>& virtual_load = std::vector<float>()) const override;
+    virtual std::vector<double> calculate_displacements(ProjectData* data, Meshing* mesh, const std::vector<double>& density = std::vector<double>(), const std::vector<double>& virtual_load = std::vector<double>()) const override;
 
     private:
-    void insert_element_matrix(std::vector<float>& K, const std::vector<float>& k, const std::vector<long>& pos, int w, int& n) const;
+    void insert_element_matrix(std::vector<double>& K, const std::vector<double>& k, const std::vector<long>& pos, int w, int& n) const;
 };
 
 }
